Move the Ammo class to its own file.
