+++ src/FieldWarning/Assets/Units/Component/Weapon/Ammo.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c3f693cf9c2fa1244aea1edb88e731f4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
